requirements.txt
IMDbPY==6.6
lxml==4.3.3
SQLAlchemy==1.3.3
XlsxWriter==1.1.8
